requirements.txt
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
Flask-Login==0.6.3
python-dotenv==1.1.0
python-docx==1.1.2
Flask-WTF==1.2.2
Pillow==11.2.1
rembg==2.0.66
requests==2.32.3
cryptography==45.0.4
num2words==0.5.14
reportlab==4.2.5
gunicorn==23.0.0
